fix conflict in alerts_service.js and its test file

+++ public-interface/dashboard/public/js/services/alerts_service.js
@@ -77,29 +77,27 @@ iotServices.factory('alertsService', ['$http', 'utilityService','sessionService'
             });
         },
         deleteAlerts: function(successCallback, errorCallback){
-            sessionService.addAccountIdPrefix('/alerts')
-            .then(function(url){
-                $http({
-                    method: 'DELETE',
-                    url: url,
-                    params: {
-                        "_" : utilityService.timeStamp()
-                    }
-                }).success(successCallback)
-                  .error(errorCallback);
-            });
+            if(confirm("Are you sure want to delete all?")){
+                sessionService.addAccountIdPrefix('/alerts')
+                .then(function(url){
+                    $http({
+                        method: 'DELETE',
+                        url: url
+                    }).success(successCallback)
+                      .error(errorCallback);
+                });
+            }
         },
         deleteAlert: function(alertId, successCallback, errorCallback){
-            sessionService.addAccountIdPrefix('/alerts/' + alertId)
-            .then(function(url){
-                $http({
-                    method: 'DELETE',
-                    url: url,
-                    params: {
-                        "_" : utilityService.timeStamp()
-                    }
-                }).success(successCallback).error(errorCallback);
-            });
+            if(confirm("Are you sure want to delete?")){
+                sessionService.addAccountIdPrefix('/alerts/' + alertId)
+                    .then(function(url) {
+                        $http({
+                            method: 'DELETE',
+                            url: url
+                        }).success(successCallback).error(errorCallback);
+                });
+            }
         },
         updateStatus: function(options, successCallback, errorCallback){
             var alert = options.alert,
@@ -153,17 +151,6 @@ iotServices.factory('alertsService', ['$http', 'utilityService','sessionService'
 
                 $http(requestOptions).success(successCallback).error(errorCallback);
             });
-        },
-        deleteAlert: function(alertId, successCallback, errorCallback){
-            if(confirm("Are you sure want to delete?")){
-                sessionService.addAccountIdPrefix('/alerts/' + alertId)
-                    .then(function(url) {
-                        $http({
-                            method: 'DELETE',
-                            url: url
-                        }).success(successCallback).error(errorCallback);
-                });
-            }
         },
         data: summary
     };
